+++ app/app/(auth)/signup.tsx
@@ -1,6 +1,6 @@
-import {Text, Modal, View, Pressable, Platform} from 'react-native';
+import {Text, View} from 'react-native';
 import React, {useState} from 'react';
-import {Controller, useForm} from 'react-hook-form';
+import {useForm} from 'react-hook-form';
 import {zodResolver} from '@hookform/resolvers/zod';
 import {SignupSchema} from '@/schema/auth.schema';
 import {ControlledTextInput} from '@/components/TextInput';
@@ -9,21 +9,14 @@ import {getColor} from '@/config/theme';
 import {ShorttieeButton} from '@/components/Button';
 import {Link, router} from 'expo-router';
 import {CustomCheckBox} from '@/components/CheckBox';
-import tw from 'twrnc';
 import {AuthScreenLayout} from '@/layouts/authLayout';
+import {GenderSelector} from '@/components/GenderSelect';
 
 type GenderT = {key: string; label: string};
-
-const genderData: GenderT[] = [
-  {key: 'male', label: 'Male'},
-  {key: 'female', label: 'Female'},
-  {key: 'others', label: 'Others'},
-];
 
 export default function Signup() {
   const [showPassword, setShowPassword] = useState<boolean>(false);
   const [isSelected, setIsSelected] = useState(false);
-  const [modalVisible, setModalVisible] = useState(false);
   const [selectedGender, setSelectedGender] = useState<GenderT>();
 
   const {
@@ -40,11 +33,6 @@ export default function Signup() {
       console.log({...data});
       router.navigate(`/(auth)/otp?email=${data.email}`);
     }
-  };
-
-  const handleSelect = (gender: GenderT) => {
-    setSelectedGender(gender);
-    setModalVisible(false);
   };
 
   return (
@@ -90,78 +78,12 @@ export default function Signup() {
               />
             }
           />
-          <Controller
+          <GenderSelector
+            selectedGender={selectedGender}
+            setSelectedGender={setSelectedGender}
             control={control}
-            name="gender"
-            render={({field}) => (
-              <View style={tw`w-full gap-1`}>
-                <Pressable
-                  style={tw`flex px-4 flex flex-row rounded-xl bg-gray-100 items-center  border border-gray-300 w-full ${Platform.select(
-                    {
-                      android: 'py-2',
-                      ios: 'py-4',
-                      default: '',
-                    },
-                  )}`}
-                  onPress={() => setModalVisible(true)}>
-                  <Feather
-                    name="user-check"
-                    size={24}
-                    color={getColor('shorttiee-grey-300')}
-                  />
-                  <Text
-                    style={tw`text-sm pl-2 flex-1 ${selectedGender ? 'text-black' : 'text-gray-400'}`}>
-                    {selectedGender ? selectedGender.label : 'Select gender'}
-                  </Text>
-                  <View>
-                    <Feather
-                      name="chevron-down"
-                      size={24}
-                      color={getColor('shorttiee-grey-300')}
-                    />
-                  </View>
-                </Pressable>
-                {errors.gender && (
-                  <Text style={tw`text-xs text-red-400 font-normal`}>
-                    Select your gender
-                  </Text>
-                )}
-
-                <Modal
-                  animationType="slide"
-                  transparent
-                  visible={modalVisible}
-                  onRequestClose={() => setModalVisible(false)}>
-                  <Pressable
-                    style={tw`flex-1 justify-center items-center bg-[#00000080]`}
-                    onPress={() => setModalVisible(false)}>
-                    <View
-                      style={tw`w-4/5 bg-white rounded-lg p-4 items-center`}>
-                      {genderData.map(gender => (
-                        <Pressable
-                          key={gender.key}
-                          onPress={() => {
-                            field.onChange(gender.key);
-                            handleSelect(gender);
-                          }}
-                          style={tw`py-3 w-full items-center`}>
-                          <View className="flex-row items-center gap-1">
-                            <Text style={tw`text-black text-base flex-1`}>
-                              {gender.label}
-                            </Text>
-                            {gender.key === selectedGender?.key && (
-                              <Feather name="check" size={18} />
-                            )}
-                          </View>
-                        </Pressable>
-                      ))}
-                    </View>
-                  </Pressable>
-                </Modal>
-              </View>
-            )}
+            errors={errors.gender}
           />
-
           <ControlledTextInput
             placeholder="Email"
             keyboardType="email-address"
